Remove bytemd
Right now all data for the proposals get saved to the chain. Therefore having extensive markdown content in our proposals will cost quite a bit of Ethers. For now, I remove the markdown editor as there are plans to use a separate discussion forum for the exact contents of a proposal.