allow actual error messages to be displayed
before always standard http error messages were displayed.
in a lot of cases this is fine, but in a fn that throws multiple
errors, it might be helpful to show the user where exactly it went
wrong.

need to encode the msg as otherwise it will end up being invalid
json and not being able to be displayed