fix auth problems, ensure routes are protected
Have to use some not so nice code to check if the user is logged
in. Depending if one does a hard reload or not, it takes a second
for the data from store to be loaded, in which case often its
detected as one is not logged in, but actually is. This new
solution prevents this with using a req to the api as fallback.