Fix negative contributions
We're seeing this issue on test, where negative contributions are saved in the database.

The main issue is that funds, that cannot be distributed today, are saved up for tomorrow, using the `future_contribution` table. The next day, if more than one person was registered for one repositories, the saved up funds from yesterday were deducted twice, as the respective code was inside a loop instead of outside.

Another issue was that `distributeFutureAdd` was not properly divided by the amount of repositories. This further contributed to the problem. If two repositories were sponsored, they both deducted the full amount of `future_contribution` from yesterday, instead of just their share.